--- app/interfaces/FullTeamStatsDTO.java
@@ -1,8 +1,0 @@
-package interfaces;
-
-import java.util.List;
-
-public class FullTeamStatsDTO {
-    private List<MatchTeamStatsDTO> matches;
-    private List<IndividualPlayerDTO> players;
-}
