+++ app/interfaces/stats/FullTeamStatsCalculationResultDTO.java
