+++ app/interfaces/stats/IndividualStandingsStatsResultDTO.java
