+++ app/interfaces/stats/RoundReportStatsCalculationResultDTO.java
